Support publishing build scan to public GE if not authenticated
Previously, if GE server was not configured, it will be automatically
configured as "ge.gradle.org". Some community members complain because
they want to publish to public GE if not authenticated. Now the logic
is changed to:

- If server url is set explicitly, use it.
- Otherwise, if current build is authenticated for ge.gradle.org, use
  it.
- Otherwise, don't set GE server url so we can publish to public GE.

This is not perfect because it doesn't take `~/.gradle/enterprise`
authentication into account, but still better than before.